requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0

# GUI
PyQt5>=5.15.0

# Network and scanning
python-nmap>=0.7.1
scapy>=2.5.0

# Report generation
jinja2>=3.1.2
markdown>=3.5.0
reportlab>=4.0.0

# Utilities
pyyaml>=6.0.1
requests>=2.31.0
colorama>=0.4.6
rich>=13.7.0

# Database
sqlalchemy>=2.0.0

# Logging
loguru>=0.7.2

# Vector database and embeddings for Phase 2
chromadb>=0.4.22
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Memory and storage
redis>=5.0.1
sqlalchemy>=2.0.0
alembic>=1.13.0

# Knowledge base
cve-search>=0.1
nvdlib>=0.7.4

# Machine learning
scikit-learn>=1.3.2
numpy>=1.24.0
pandas>=2.1.0

# Pattern recognition
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# Enhanced analysis
networkx>=3.2.1
matplotlib>=3.8.2


# Phase 2 & 3 Dependencies
# Add to requirements.txt

# Vector Database (Phase 2)
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Enhanced ML/AI (Phase 2)
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Redis for agent coordination (Phase 3)
redis>=5.0.0
celery>=5.3.0

# Async/Parallel Processing (Phase 3)
asyncio>=3.4.3
aiohttp>=3.9.0
multiprocessing-logging>=0.3.0


# Phase 4: Advanced Evasion & Stealth
# Additional dependencies
scikit-learn>=1.0.0


# Phase 5: Zero-Day Discovery & Exploit Development Requirements

# Fuzzing Tools
afl-utils>=1.35  # AFL++ utilities
honggfuzz>=2.5  # Honggfuzz fuzzer (if available via pip)

# Symbolic Execution
angr>=9.2.0  # Binary analysis and symbolic execution
claripy>=9.2.0  # Constraint solver for angr
pyvex>=9.2.0  # VEX IR translation

# Binary Analysis
r2pipe>=1.7.0  # Radare2 Python bindings
capstone>=5.0.0  # Disassembly framework
keystone-engine>=0.9.2  # Assembler framework
unicorn>=2.0.0  # CPU emulator

# Reverse Engineering
pyelftools>=0.29  # ELF file parsing
pefile>=2023.2.7  # PE file parsing
lief>=0.13.0  # Binary parsing (ELF, PE, Mach-O)

# Exploit Development
pwntools>=4.10.0  # Exploit development framework
ropper>=1.13.8  # ROP gadget finder
ROPgadget>=7.3  # Alternative ROP gadget finder

# Code Analysis
tree-sitter>=0.20.0  # Fast code parsing
libclang>=16.0.0  # C/C++ parsing
jedi>=0.19.0  # Python code analysis
astroid>=3.0.0  # Python AST analysis

# Static Analysis
bandit>=1.7.5  # Python security analysis
pylint>=3.0.0  # Python linting
semgrep>=1.45.0  # Multi-language static analysis

# Dynamic Analysis (Python bindings where available)
# Note: Many require system installation
# - Valgrind (system package)
# - PIN (Intel binary instrumentation)
# - DynamoRIO (binary instrumentation)
# - Frida (dynamic instrumentation)
frida>=16.1.0
frida-tools>=12.2.0

# Cryptography Analysis
cryptography>=41.0.0  # Crypto library for testing
pycryptodome>=3.19.0  # Crypto primitives

# Network/API Testing
aiohttp>=3.9.0  # Async HTTP client
httpx>=0.25.0  # HTTP client
requests>=2.31.0  # HTTP library

# Data Processing
lxml>=4.9.3  # XML processing
pyyaml>=6.0.1  # YAML processing
msgpack>=1.0.7  # MessagePack serialization

# Machine Learning (for pattern recognition)
scikit-learn>=1.3.0  # ML algorithms
numpy>=1.24.0  # Numerical computing
pandas>=2.0.0  # Data analysis

# Utilities
tqdm>=4.66.0  # Progress bars
colorama>=0.4.6  # Terminal colors
rich>=13.7.0  # Rich terminal output
psutil>=5.9.0  # Process and system utilities

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.12.0
isort>=5.13.0
mypy>=1.7.0


# Phase 6: Advanced Persistence & Command Infrastructure Requirements

# Core async libraries
aiohttp>=3.8.0
aiofiles>=23.0.0
aiodns>=3.0.0

# Cryptography
pycryptodome>=3.18.0
cryptography>=41.0.0

# Kerberos and Windows authentication
pypykatz>=0.6.0
impacket>=0.11.0
pywin32>=306; platform_system=="Windows"
wmi>=1.5.1; platform_system=="Windows"

# Network and protocol handling
scapy>=2.5.0
dnspython>=2.4.0
websockets>=11.0

# Cloud providers
boto3>=1.28.0  # AWS
azure-storage-blob>=12.17.0  # Azure
google-cloud-storage>=2.10.0  # Google Cloud
dropbox>=11.36.0  # Dropbox
requests>=2.31.0  # General HTTP

# Image processing (for steganography)
Pillow>=10.0.0
numpy>=1.24.0

# Database access
PyMySQL>=1.1.0
psycopg2-binary>=2.9.0
pymongo>=4.4.0

# SSH/FTP clients
paramiko>=3.3.0
pysftp>=0.2.9

# Browser automation
selenium>=4.11.0
pyppeteer>=1.0.2

# Linux-specific
python-pam>=2.0.2; platform_system=="Linux"
evdev>=1.6.1; platform_system=="Linux"

# Password cracking
hashcat-utils>=1.9.0
passlib>=1.7.4

# Persistence and rootkit tools
pefile>=2023.2.7  # PE file manipulation
pyelftools>=0.29  # ELF file manipulation

# Memory analysis
volatility3>=2.4.0

# Additional utilities
python-dotenv>=1.0.0
pyyaml>=6.0
colorama>=0.4.6
tabulate>=0.9.0
tqdm>=4.66.0

# Already required from previous phases
# (include if running Phase 6 standalone)
openai>=1.0.0
anthropic>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0


# Phase 7: Lateral Movement & Domain Dominance Requirements

# Active Directory Tools
impacket>=0.11.0
neo4j>=5.14.0
ldap3>=2.9.1
pyasn1>=0.5.0

# Network Tools
paramiko>=3.3.1
pexpect>=4.8.0
python-nmap>=0.7.1
scapy>=2.5.0

# Database Clients
pymssql>=2.2.8
psycopg2-binary>=2.9.9
pymongo>=4.5.0
mysql-connector-python>=8.2.0

# Container/Cloud
docker>=6.1.3
kubernetes>=28.1.0
boto3>=1.29.7
azure-identity>=1.14.1
google-cloud-compute>=1.15.0

# Cryptography
cryptography>=41.0.7
pycryptodome>=3.19.0

# HTTP/Web
aiohttp>=3.9.1
requests>=2.31.0

# Process/System
psutil>=5.9.6
pywin32>=306; platform_system == "Windows"

# Networking
dnspython>=2.4.2
netaddr>=0.9.0

# Privilege Escalation
pefile>=2023.2.7; platform_system == "Windows"

# Utilities
python-dateutil>=2.8.2
pyyaml>=6.0.1
colorama>=0.4.6


# Phase 8: Data Exfiltration & Impact Analysis Requirements

# Data Discovery
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'
Pillow>=10.0.0
PyPDF2>=3.0.0

# DNS and Network
dnspython>=2.4.0
aiohttp>=3.9.0

# Database Analysis
pymysql>=1.1.0
psycopg2-binary>=2.9.9

# Steganography
numpy>=1.24.0

# All previous phases
-r requirements-phase7.txt
